+++ Extensions/MyCommonConstants.cs
@@ -1,6 +1,6 @@
 ﻿namespace MyBox
 {
-	public class MyCommonConstants
+	public static class MyCommonConstants
 	{
 		public static readonly System.Random SystemRandom = new System.Random();
 	}